HW16. Editing, saving, deleting and viewing of resumes in CRUD added; and all functions work

+++ src/ru/javawebinar/basejava/web/ResumeServlet.java
@@ -1,8 +1,7 @@
 package ru.javawebinar.basejava.web;
 
 import ru.javawebinar.basejava.Config;
-import ru.javawebinar.basejava.model.ContactType;
-import ru.javawebinar.basejava.model.Resume;
+import ru.javawebinar.basejava.model.*;
 import ru.javawebinar.basejava.storage.SqlStorage;
 
 import javax.servlet.ServletException;
@@ -10,6 +9,8 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.util.Arrays;
+import java.util.List;
 
 public class ResumeServlet extends HttpServlet {
     private SqlStorage sqlStorage;
@@ -25,17 +26,40 @@ public class ResumeServlet extends HttpServlet {
         request.setCharacterEncoding("UTF-8");
         String uuid = request.getParameter("uuid");
         String fullName = request.getParameter("fullName");
-        Resume r = sqlStorage.get(uuid);
-        r.setFullName(fullName);
-        for (ContactType type : ContactType.values()) {
-            String value = request.getParameter(type.name());
-            if (value != null && value.trim().length() != 0) {
-                r.setContacts(type, value);
-            } else {
-                r.getContacts().remove(type);
+        Resume resume;
+
+        if (isEmpty(uuid)) {
+            resume = new Resume(fullName);
+        } else {
+            resume = sqlStorage.get(uuid);
+            resume.setFullName(fullName);
+        }
+
+        for (ContactType ct : ContactType.values()) {
+            String value = request.getParameter(ct.name());
+            resume.setContacts(ct, value);
+        }
+
+        for (SectionType st : SectionType.values()) {
+            String value = request.getParameter(st.name());
+            switch (st) {
+                case PERSONAL:
+                case OBJECTIVE:
+                    resume.setSections(st, new TextSection(value));
+                    break;
+                case ACHIEVEMENT:
+                case QUALIFICATION:
+                    List<String> skills = Arrays.asList(value.split("\n"));
+                    resume.setSections(st, new ListSection(skills));
+                    break;
             }
         }
-        sqlStorage.update(r);
+
+        if (isEmpty(uuid)) {
+            sqlStorage.save(new Resume(fullName));
+        } else {
+            sqlStorage.update(resume);
+        }
         response.sendRedirect("resume");
     }
 
@@ -49,27 +73,36 @@ public class ResumeServlet extends HttpServlet {
         }
 
         Resume resume = null;
+        String page;
 
         switch (action) {
+            case "submitClear":
+            case "submitDelete":
+                page = "/WEB-INF/jsp/clear.jsp";
+                break;
             case "clear":
-                sqlStorage.clear();
-                response.sendRedirect("resume");
-                return;
             case "delete":
-                sqlStorage.delete(uuid);
+                if (action.equals("clear")) {
+                    sqlStorage.clear();
+                } else {
+                    sqlStorage.delete(uuid);
+                }
                 response.sendRedirect("resume");
                 return;
+            case "add":
             case "view":
             case "edit":
-                resume = sqlStorage.get(uuid);
+                resume = action.equals("add") ? new Resume() : sqlStorage.get(uuid);
+                page = action.equals("view") ? "/WEB-INF/jsp/view.jsp" : "/WEB-INF/jsp/edit.jsp";
                 break;
             default:
                 throw new IllegalArgumentException("Action " + action + " is illegal");
         }
         request.setAttribute("resume", resume);
-        request.getRequestDispatcher(
-                ("view".equals(action) ? "/WEB-INF/jsp/view.jsp" : "/WEB-INF/jsp/edit.jsp")
-        ).forward(request, response);
+        request.getRequestDispatcher(page).forward(request, response);
+    }
 
+    private boolean isEmpty(String value) {
+        return value == null || value.trim().length() == 0;
     }
 }
